app/requirements.txt
requests
beautifulsoup4
pymongo
langdetect
pandas
python-dotenv